requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
funcsigs==1.0.2
html5lib==0.9999999
Markdown==2.2.0
mock==2.0.0
numpy==1.13.0
olefile==0.44
pbr==3.0.1
Pillow==4.1.1
protobuf==3.3.0
scipy==0.19.0
six==1.10.0
tensorflow-gpu==1.2.0
Werkzeug==0.12.2
tqdm==4.11.2
imageio==2.1.2
decorator==4.0.11
moviepy==0.2.3.2